+++ crates/icu/src/interface/response.rs
@@ -2,10 +2,10 @@ use crate::{
     Error,
     ic::api::{canister_cycle_balance, msg_caller},
     interface::{
-        ic::{deposit_cycles, ic_create_canister_full, ic_upgrade_canister},
+        ic::{create_canister_full, deposit_cycles, upgrade_canister},
         request::{CreateCanisterRequest, CyclesRequest, Request, UpgradeCanisterRequest},
     },
-    memory::CanisterState,
+    memory::canister_state::CanisterState,
     state::canister::CanisterRegistry,
 };
 use candid::{CandidType, Principal};
@@ -63,7 +63,7 @@ pub async fn response(req: Request) -> Result<Response, Error> {
 // create_canister_response
 async fn create_canister_response(req: &CreateCanisterRequest) -> Result<Response, Error> {
     let new_canister_pid =
-        ic_create_canister_full(&req.canister_type, &req.parents, req.extra_arg.clone()).await?;
+        create_canister_full(&req.canister_type, &req.parents, req.extra_arg.clone()).await?;
 
     Ok(Response::CreateCanister(CreateCanisterResponse {
         new_canister_pid,
@@ -73,7 +73,7 @@ async fn create_canister_response(req: &CreateCanisterRequest) -> Result<Respons
 // upgrade_canister_response
 async fn upgrade_canister_response(req: &UpgradeCanisterRequest) -> Result<Response, Error> {
     let canister = CanisterRegistry::try_get(&req.canister_type)?;
-    ic_upgrade_canister(req.canister_pid, canister.wasm).await?;
+    upgrade_canister(req.canister_pid, canister.wasm).await?;
 
     Ok(Response::UpgradeCanister(UpgradeCanisterResponse {}))
 }